Add RenameRepositoryInPlace to RepositoryStore
This commit adds RenameRepositoryInPlace to RepositoryStore. The new method
will replace RenameRepository which can be removed in a later release.
Unlike RenameRepository, renames doesn't change the replica path stored in
the database. As Praefect is now routing using the replica_path, this enables
Praefect to rename repositories atomically in the database without necessiating
any disk changes. The old RenameRepository is left in place so any in-flight rename
jobs can still be processed by it in 14.6.