Resolve replica path in TestRestoreSubcommand
TestRestoreSubcommand verifies after restoring a backup that everything
is good by peeking into the filesystem and checking the repository. The
test currently assumes that the RelativePath is the actual storage path
of the repository. This is not the case if the test case runs with Praefect
as Praefect would generate a unique path for the repository. The test
case thus fails when ran with Praefect when Praefect rewrites the relative
path. This commit updates the test case to resolve the replica path from
Praefect if the test case is ran with Praefect.