cmd/gitaly: Deduplicate creation of command factories
The "gitaly" executable is executing two different Git command
factories, once to verify that we got a supported Git version and once
as a dependency for the server's setup. The factory is about to
introduce holding state, which would thus also be created twice without
much of a reason.

Refactor the code to only construct a single Git command factory and
reuse it for both cases.