datastore: Add metric to keep track of replication queue depth
Relication queue depth is something that is very useful when debugging
slowness in Praefect. There was an incident
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/6133
where the replication queue depth would have been very useful to have.

This commit adds the metric to the datastore collector. On the
replication_queue table there is an index on virtual_storage and
target_node_storage on the job jsonb column, so this query does use an
index scan.

Changelog: added