cmd/gitaly-wrapper: Refactor reading of PID file to be testable
It's hard to test reading the logic which reads the PID file because it
depends on environment variables to be set. Refactor the function to
instead accept the PID file path as parameter and add a bunch of tests
to verify it works as expected.

While at it, renames the function to `readPIDFile()` to more clearly
show what it does.