cmd/gitaly-wrapper: Add tests for finding processes by PID file
Now that `readPIDFile()` doesn't depend on global state anymore it's a
lot easier to test finding processes via such a PID file. Add a bunch of
tests to verify it works as expected.

While at it, rename the function to `findProcess()` instead of
`findGitaly()`. The executable is used for both Praefect and Gitaly, so
the found process may not in fact be a Gitaly instance.