command: Fix log message verification
The root cause of the change is a flaky behaviour
of a couple of the tests in the command package.
The tests assume the first log message would be
the output of the bash script that is invoked.
But it happens that 'spawn token acquired' text
is the first message instead. It seems weird why
it differs from time to time, but the solution is
pretty strait forward: search for th message we
are interested in. All others will be skipped.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3980