sidechannel: proxy: allow early upstream return
Before this change, it was a proxy error if a sidechannel upstream
closed its connection without reading from downstream until EOF first.
But it turns out we need to support that use case.

This commit changes the proxy control flow so that it is not an error
if the upstream closes its connection before the downstream has been
fully consumed.

Changelog: fixed