git: Fix permissions of binary directory
One of our tests is creating the binary directory with permissions 644,
which means that it cannot be searched by default. This didn't surface
in any failing tests, first because CI was running as root user and
second because we only to write something into that directory in case
tests are running with bundlde Git. Now that we do execute CI as
unprivileged user though the test fails as expected.

Fix this bug by creating the directory with 755 permissions.